Agentic UAE Expansion Advisor/requirements.txt
agno>=0.1.0               # Agentic AI framework
openai>=1.0.0             # OpenAI API client
exa-py>=0.1.0             # Exa search API client
python-dotenv>=1.0.0      # For loading .env API keys
playwright>=1.40.0        # For web scraping
beautifulsoup4>=4.12.0    # HTML parsing
lxml>=4.9.2               # Parser for BeautifulSoup